CompreFace_scripts/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
compreface-sdk==0.6.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.3
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
# requests-toolbelt==0.10.1
seaborn==0.13.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
